refactor: extract aceclaw-learning batch 2 — skill / validation / maintenance sub-packages
Continues the daemon decomposition started in #503. Daemon was at 28,295 LoC
after batch 1; this batch moves another ~5K LoC out into themed sub-packages
under aceclaw-learning, leaving daemon at ~23,000 LoC.

## What moved

16 source files (+12 tests) into 3 new sub-packages under aceclaw-learning:

  dev.aceclaw.learning.skill          (6 files, 1303 LoC)
    SkillDraftEvent, SkillDraftEventFeed, SkillDraftGenerator
    SkillMemoryFeedback, SkillMetricsStore, SkillRefinementEngine

  dev.aceclaw.learning.validation     (5 files, 1585 LoC)
    LearningValidation, LearningValidationStore, LearningValidationRecorder
    ValidationGateEngine, AutoReleaseController
    (AutoReleaseController grouped here since it's the action-side of
     validation — keeps the validation pipeline self-contained.)

  dev.aceclaw.learning.maintenance    (5 files, 843 LoC)
    LearningMaintenanceCandidateBridge, LearningMaintenanceRecoveryStore
    LearningMaintenanceRun, LearningMaintenanceRunStore
    LearningMaintenanceScheduler

Total: 16 files + 12 tests moved, daemon drops ~3,731 LoC of code that had
no business being in the same module as the UDS listener / JSON-RPC router.

## What did NOT move (and why)

Three files have hard daemon-internal coupling and are deferred:

  - DynamicSkillGenerator       (851 LoC) → references AgentSession
  - SessionSkillPacker          (435 LoC) → references AgentSession,
                                            SessionHistoryStore, SessionManager
  - SessionSkillPackPrompt      (297 LoC) → references AgentSession

These need AgentSession either inlined or extracted to core first. Out of
scope for this PR; queued for a later batch with the session subsystem.

Two existing daemon-test files stayed in daemon test source because they
depend on MockLlmClient (which lives in daemon test):
  - SkillRefinementEngineTest, SkillDraftGeneratorTest
Both updated with cross-package imports to point at the moved source classes.
The source classes themselves moved cleanly. Unifying MockLlmClient into a
shared test-fixture module is a separate cleanup.

## Visibility bumps

Several methods/records in moved classes were package-private and worked
when daemon and these classes lived in the same package. Now that they
cross a module boundary, the few that StreamingAgentHandler and the moved-
back tests need had to be bumped to public:

  SkillRefinementEngine:
    - methods: analyze, prepare, proposeRefinement, apply
    - records: SkillRefinement, PreparedPlan, WindowStats, RefinementState
    - inner: RefinementOutcome.none() static factory

  ValidationGateEngine:
    - secondary constructor (Clock, boolean, ...) used by tests for time
      injection

These are the same APIs daemon was already using; visibility was an
artifact of same-package, not an intentional encapsulation decision.

## Consumer imports added

  aceclaw-daemon (3 files):
    AceClawDaemon              — 13 imports (all 3 sub-packages)
    DynamicSkillGenerator      — 2 validation imports
    StreamingAgentHandler      — 3 skill + 2 validation imports

  aceclaw-daemon tests (4 files):
    CandidatePipelineIntegrationTest, DynamicSkillGeneratorTest,
    SkillDraftGeneratorTest, SkillRefinementEngineTest

Inside aceclaw-learning, one cross-subpackage import:
    validation/LearningValidationRecorder → skill/SkillRefinementEngine

## Test plan

- [x] :aceclaw-learning:test — 68 tests pass across 14 test classes
  (14 ErrorDetector + 9 FailureSignal + 2 each for batch-1 stores
  + 6 skill/validation/maintenance suites)
- [x] All daemon consumer tests pass:
    SelfImprovementEngineTest (24), DynamicSkillGeneratorTest (13),
    SkillRefinementEngineTest (4), SkillDraftGeneratorTest (5),
    StreamingAgentHandlerPostLearningTest (1),
    CandidatePipelineIntegrationTest (10)
- [x] Pre-existing flaky tests verified flaky on main (NOT introduced here):
    SelfLearningPipelineTest (2 failures), SessionSkillPackerTest (1 failure)
- [x] Full compile across 15 modules

## Stats

Diff is +68 / -38 across 33 files — git rename detection sees every move
correctly. Daemon goes from 28,295 → ~23,000 LoC (a further 19% reduction
on top of batch 1's 3.3%).

## Remaining

Batch 3+ targets, when the session subsystem gets carved out:
  - DynamicSkillGenerator + SessionSkillPacker + SessionSkillPackPrompt
    (move with AgentSession resolution)
  - PatternDetector (also needs AgentSession)
  - Possibly: SessionAnalyzer / SessionHistoryStore / SessionEndExtractor
    bundled together as the learning "session view" layer
