+++ aceclaw-daemon/src/main/java/dev/aceclaw/daemon/AceClawDaemon.java
@@ -28,6 +28,19 @@ import dev.aceclaw.learning.LearningExplanationRecorder;
 import dev.aceclaw.learning.LearningExplanationStore;
 import dev.aceclaw.learning.LearningSignalReview;
 import dev.aceclaw.learning.LearningSignalReviewStore;
+import dev.aceclaw.learning.maintenance.LearningMaintenanceCandidateBridge;
+import dev.aceclaw.learning.maintenance.LearningMaintenanceRecoveryStore;
+import dev.aceclaw.learning.maintenance.LearningMaintenanceRun;
+import dev.aceclaw.learning.maintenance.LearningMaintenanceRunStore;
+import dev.aceclaw.learning.maintenance.LearningMaintenanceScheduler;
+import dev.aceclaw.learning.skill.SkillDraftEvent;
+import dev.aceclaw.learning.skill.SkillDraftEventFeed;
+import dev.aceclaw.learning.skill.SkillDraftGenerator;
+import dev.aceclaw.learning.validation.AutoReleaseController;
+import dev.aceclaw.learning.validation.LearningValidation;
+import dev.aceclaw.learning.validation.LearningValidationRecorder;
+import dev.aceclaw.learning.validation.LearningValidationStore;
+import dev.aceclaw.learning.validation.ValidationGateEngine;
 import dev.aceclaw.llm.LlmClientFactory;
 import dev.aceclaw.memory.AutoMemoryStore;
 import dev.aceclaw.memory.CandidateStateMachine;
