+++ aceclaw-daemon/src/main/java/dev/aceclaw/daemon/StreamingAgentHandler.java
@@ -40,6 +40,11 @@ import dev.aceclaw.core.planner.PlanExecutionResult;
 import dev.aceclaw.core.planner.PlanStatus;
 import dev.aceclaw.learning.LearningExplanation;
 import dev.aceclaw.learning.LearningExplanationRecorder;
+import dev.aceclaw.learning.skill.SkillMemoryFeedback;
+import dev.aceclaw.learning.skill.SkillMetricsStore;
+import dev.aceclaw.learning.skill.SkillRefinementEngine;
+import dev.aceclaw.learning.validation.LearningValidation;
+import dev.aceclaw.learning.validation.LearningValidationRecorder;
 import dev.aceclaw.core.planner.PlannedStep;
 import dev.aceclaw.core.planner.SequentialPlanExecutor;
 import dev.aceclaw.core.planner.StepResult;
