+++ aceclaw-daemon/src/test/java/dev/aceclaw/daemon/DynamicSkillGeneratorTest.java
@@ -10,6 +10,9 @@ import dev.aceclaw.core.llm.Usage;
 import dev.aceclaw.learning.LearningExplanation;
 import dev.aceclaw.learning.LearningExplanationRecorder;
 import dev.aceclaw.learning.LearningExplanationStore;
+import dev.aceclaw.learning.validation.LearningValidation;
+import dev.aceclaw.learning.validation.LearningValidationRecorder;
+import dev.aceclaw.learning.validation.LearningValidationStore;
 import dev.aceclaw.memory.Insight;
 import dev.aceclaw.memory.PatternType;
 import org.junit.jupiter.api.BeforeEach;
