+++ aceclaw-learning/src/main/java/dev/aceclaw/learning/maintenance/LearningMaintenanceRecoveryStore.java
@@ -1,4 +1,4 @@
-package dev.aceclaw.daemon;
+package dev.aceclaw.learning.maintenance;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
