+++ aceclaw-learning/src/main/java/dev/aceclaw/learning/maintenance/LearningMaintenanceRun.java
@@ -1,4 +1,4 @@
-package dev.aceclaw.daemon;
+package dev.aceclaw.learning.maintenance;
 
 import java.time.Instant;
 import java.util.Objects;
