+++ aceclaw-learning/src/main/java/dev/aceclaw/learning/maintenance/LearningMaintenanceScheduler.java
@@ -1,4 +1,4 @@
-package dev.aceclaw.daemon;
+package dev.aceclaw.learning.maintenance;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
