+++ aceclaw-learning/src/main/java/dev/aceclaw/learning/skill/SkillDraftEventFeed.java
@@ -1,4 +1,4 @@
-package dev.aceclaw.daemon;
+package dev.aceclaw.learning.skill;
 
 import java.util.ArrayDeque;
 import java.util.ArrayList;
