+++ aceclaw-learning/src/main/java/dev/aceclaw/learning/validation/LearningValidation.java
@@ -1,4 +1,4 @@
-package dev.aceclaw.daemon;
+package dev.aceclaw.learning.validation;
 
 import java.time.Instant;
 import java.util.List;
