+++ aceclaw-learning/src/test/java/dev/aceclaw/learning/validation/AutoReleaseControllerTest.java
@@ -1,4 +1,4 @@
-package dev.aceclaw.daemon;
+package dev.aceclaw.learning.validation;
 
 import dev.aceclaw.memory.CandidateKind;
 import dev.aceclaw.memory.CandidateState;
