+++ aceclaw-learning/src/test/java/dev/aceclaw/learning/validation/LearningValidationStoreTest.java
@@ -1,4 +1,4 @@
-package dev.aceclaw.daemon;
+package dev.aceclaw.learning.validation;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
